+++ src/main/java/com/mapera/inventory_system/domain/entity/Category.java
@@ -1,15 +1,15 @@
 package com.mapera.inventory_system.domain.entity;
 
 public class Category {
-    private final int id;
+    private final Long id;
     private String name;
 
-    public Category(int id, String name) {
+    public Category(Long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public int getId() {
+    public Long getId() {
         return this.id;
     }
